Share more code between FSD versions
