+++ Inferences/ForwardSubsumptionDemodulation.cpp
@@ -52,122 +52,6 @@ void ForwardSubsumptionDemodulation::detach()
 }
 
 
-namespace {
-
-
-/**
- * A binder that consists of two maps: a base and an overlay.
- * Lookup first checks the base map, then the overlay map.
- * New bindings are added to the overlay map.
- *
- * In FSD, the base bindings are extracted from the MLMatcher and correspond to the subsumption part,
- * while the overlay bindings are from the demodulation part (i.e., from
- * matching the lhs of the demodulation equality to the candidate term that
- * might be simplified).
- *
- * This class implements the Binder interface as described in Kernel/Matcher.hpp,
- * and the Applicator interface as described in Kernel/SubstHelper.hpp.
- */
-class OverlayBinder
-{
-  public:
-    using Var = unsigned int;
-    using BindingsMap = v_unordered_map<Var, TermList>;
-
-    OverlayBinder()
-      : m_base(16)
-      , m_overlay(16)
-    { }
-
-    /// Initializes the base bindings with the given argument
-    explicit
-    OverlayBinder(BindingsMap&& initialBindings)
-      : m_base(std::move(initialBindings))
-      , m_overlay(16)
-    { }
-
-    bool bind(Var var, TermList term)
-    {
-      // If the variable is already bound, it must be bound to the same term.
-      auto base_it = m_base.find(var);
-      if (base_it != m_base.end()) {
-        return base_it->second == term;
-      }
-      else {
-        auto res = m_overlay.insert({var, term});
-        auto it = res.first;
-        bool inserted = res.second;
-        return inserted || (it->second == term);
-      }
-    }
-
-    bool isBound(Var var)
-    {
-      return m_base.find(var) != m_base.end()
-        || m_overlay.find(var) != m_overlay.end();
-    }
-
-    void specVar(Var var, TermList term)
-    {
-      ASSERTION_VIOLATION;
-    }
-
-    /// Clear all bindings
-    void clear()
-    {
-      m_base.clear();
-      m_overlay.clear();
-    }
-
-    /// Direct access to base bindings.
-    /// The returned map may only be modified if the overlay map is empty!
-    /// (this function is unfortunately necessary to be able to extract the base bindings from the MLMatcher without dynamic memory allocation)
-    BindingsMap& base()
-    {
-      ASS(m_overlay.empty());
-      return m_base;
-    }
-
-    /// Resets to base bindings
-    void reset() {
-      m_overlay.clear();
-    }
-
-    // Makes objects of this class work as applicator for substitution
-    // (as defined in Kernel/SubstHelper.hpp)
-    TermList apply(Var var) const {
-      auto b_it = m_base.find(var);
-      if (b_it != m_base.end()) {
-        return b_it->second;
-      } else {
-        auto o_it = m_overlay.find(var);
-        if (o_it != m_overlay.end()) {
-          return o_it->second;
-        } else {
-          // We should never access unbound variables
-          // (NOTE: we should not return the variable itself here, as this creates a risk of mixing variables coming from different clauses)
-          ASSERTION_VIOLATION;
-        }
-      }
-    }
-
-    TermList applyTo(TermList t, bool noSharing = false) const {
-      return SubstHelper::apply(t, *this, noSharing);
-    }
-
-    Literal* applyTo(Literal* l) const {
-      return SubstHelper::apply(l, *this);
-    }
-
-  private:
-    BindingsMap m_base;
-    BindingsMap m_overlay;
-};  // class OverlayBinder
-
-
-}  // namespace
-
-
 #define CHECK_FOR_MULTIPLE_RESULTS 0
 
 bool ForwardSubsumptionDemodulation::perform(Clause* cl, Clause*& replacement, ClauseIterator& premises)
